refactor: VideoRenderer component and updatesubscriptions
