+++ packages/react-native-sdk/src/components/call/Lobby.tsx
@@ -10,11 +10,11 @@ import {
 import { CallControlsButton } from '../utility/internal/CallControlsButton';
 import { theme } from '../../theme';
 import { useLocalVideoStream } from '../../hooks';
-import { VideoRenderer } from '../utility/internal/VideoRenderer';
 import { Avatar } from '../utility/Avatar';
 import { StreamVideoParticipant } from '@stream-io/video-client';
 import { LOCAL_VIDEO_VIEW_STYLE } from '../../constants';
 import { useMediaStreamManagement } from '../../providers/MediaStreamManagement';
+import { RTCView } from 'react-native-webrtc';
 
 /**
  * Use this view prior to joining a call.
@@ -90,9 +90,9 @@ export const Lobby = ({ joinCallButton }: LobbyProps) => {
             <View style={styles.videoView}>
               <View style={styles.topView} />
               {isVideoAvailable ? (
-                <VideoRenderer
+                <RTCView
                   mirror={isCameraOnFrontFacingMode}
-                  mediaStream={localVideoStream}
+                  streamURL={localVideoStream?.toURL()}
                   objectFit="cover"
                   style={StyleSheet.absoluteFillObject}
                 />
