+++ packages/react-native-sdk/src/components/participants/ParticipantVideoFallback.tsx
@@ -1,31 +1,36 @@
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import { LayoutChangeEvent, StyleSheet, Text, View } from 'react-native';
 import { Avatar } from '../utility';
 import { StreamVideoParticipant } from '@stream-io/video-client';
 import { theme } from '../../theme';
 
-export type ParticipantVideoPlaceholderProps = {
+export type ParticipantVideoFallbackProps = {
+  /**
+   * The participant whose info will be displayed.
+   */
   participant: StreamVideoParticipant;
+  /**
+   * Invoked on mount and layout changes with
+   * {nativeEvent: { layout: {x, y, width, height}}}.
+   */
+  onLayout?: ((event: LayoutChangeEvent) => void) | undefined;
 };
 
-export const ParticipantVideoPlaceholder = ({
+export const ParticipantVideoFallback = ({
   participant,
-}: ParticipantVideoPlaceholderProps) => {
+  onLayout,
+}: ParticipantVideoFallbackProps) => {
   const { name, image, userId } = participant;
   const participantLabel = name ?? userId;
 
   // Display the Participant name/user id if the image isn't present.
-  if (!image) {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.label}>{participantLabel}</Text>
-      </View>
-    );
-  }
-
   return (
-    <View style={styles.container}>
-      <Avatar participant={participant} />
+    <View style={styles.container} onLayout={onLayout}>
+      {!image ? (
+        <Text style={styles.label}>{participantLabel}</Text>
+      ) : (
+        <Avatar participant={participant} />
+      )}
     </View>
   );
 };
