+++ packages/react-native-sdk/src/components/participants/VideoRenderer.tsx
@@ -1,0 +1,209 @@
+import React, { useEffect, useRef } from 'react';
+import { StyleSheet } from 'react-native';
+import { RTCView } from 'react-native-webrtc';
+import { ParticipantVideoType, ParticipantViewProps } from '.';
+import {
+  CallingState,
+  SfuModels,
+  StreamVideoParticipant,
+  VisibilityState,
+} from '@stream-io/video-client';
+import { useCall, useCallCallingState } from '@stream-io/video-react-bindings';
+import { useMediaStreamManagement } from '../../providers';
+import { Z_INDEX } from '../../constants';
+import { ParticipantVideoFallback as DefaultParticipantVideoFallback } from './ParticipantVideoFallback';
+
+/**
+ * Props to be passed for the VideoRenderer component.
+ */
+export type VideoRendererProps = Pick<
+  ParticipantViewProps,
+  'ParticipantVideoFallback'
+> & {
+  /**
+   * The video kind that will be displayed.
+   * @types `screen` or `video`
+   */
+  videoMode: ParticipantVideoType;
+  /**
+   * The participant whose info will be displayed.
+   */
+  participant: StreamVideoParticipant;
+  /**
+   * When set to false, the video stream will not be displayed even if it is available.
+   *
+   * @default false
+   */
+  muteVideo?: boolean;
+};
+
+/**
+ * Lower level component, that represents only the video part (wrapper around the WebRTC)
+ */
+export const VideoRenderer = ({
+  videoMode,
+  participant,
+  muteVideo,
+  ParticipantVideoFallback = DefaultParticipantVideoFallback,
+}: VideoRendererProps) => {
+  const call = useCall();
+  const callingState = useCallCallingState();
+  const pendingVideoLayoutRef = useRef<SfuModels.VideoDimension>();
+  const subscribedVideoLayoutRef = useRef<SfuModels.VideoDimension>();
+  const { isCameraOnFrontFacingMode } = useMediaStreamManagement();
+  const {
+    isLocalParticipant,
+    sessionId,
+    publishedTracks,
+    viewportVisibilityState,
+    videoStream,
+    screenShareStream,
+  } = participant;
+
+  const isScreenSharing = videoMode === 'screen';
+  const isPublishingVideoTrack = publishedTracks.includes(
+    isScreenSharing
+      ? SfuModels.TrackType.SCREEN_SHARE
+      : SfuModels.TrackType.VIDEO,
+  );
+  const hasJoinedCall = callingState === CallingState.JOINED;
+  const canShowVideo = !!videoStream && !muteVideo && isPublishingVideoTrack;
+  const videoStreamToRender = isScreenSharing ? screenShareStream : videoStream;
+  const mirror = isLocalParticipant && isCameraOnFrontFacingMode;
+
+  /**
+   * This effect updates the participant's viewportVisibilityState
+   * Additionally makes sure that when this view becomes visible again, the layout to subscribe is known
+   */
+  useEffect(() => {
+    if (!call) {
+      return;
+    }
+    if (!muteVideo) {
+      if (viewportVisibilityState !== VisibilityState.VISIBLE) {
+        call.state.updateParticipant(sessionId, (p) => ({
+          ...p,
+          viewportVisibilityState: VisibilityState.VISIBLE,
+        }));
+      }
+    } else {
+      if (viewportVisibilityState !== VisibilityState.INVISIBLE) {
+        call.state.updateParticipant(sessionId, (p) => ({
+          ...p,
+          viewportVisibilityState: VisibilityState.INVISIBLE,
+        }));
+      }
+      if (subscribedVideoLayoutRef.current) {
+        // when video is enabled again, we want to use the last subscribed dimension to resubscribe
+        pendingVideoLayoutRef.current = subscribedVideoLayoutRef.current;
+        subscribedVideoLayoutRef.current = undefined;
+      }
+    }
+  }, [sessionId, viewportVisibilityState, muteVideo, call]);
+
+  useEffect(() => {
+    if (!hasJoinedCall && subscribedVideoLayoutRef.current) {
+      // when call is joined again, we want to use the last subscribed dimension to resubscribe
+      pendingVideoLayoutRef.current = subscribedVideoLayoutRef.current;
+      subscribedVideoLayoutRef.current = undefined;
+    }
+  }, [hasJoinedCall]);
+
+  /**
+   * This effect updates the subscription either
+   * 1. when video tracks are published and was unpublished before
+   * 2. when the view's visibility changes
+   * 3. when call was rejoined
+   */
+  useEffect(() => {
+    // NOTE: We only want to update the subscription if the pendingVideoLayoutRef is set
+    const updateIsNeeded = pendingVideoLayoutRef.current;
+
+    if (!updateIsNeeded || !call || !isPublishingVideoTrack || !hasJoinedCall) {
+      return;
+    }
+
+    // NOTE: When the view is not visible, we want to subscribe to audio only.
+    // We unsubscribe their video by setting the dimension to undefined
+    const dimension = !muteVideo ? pendingVideoLayoutRef.current : undefined;
+
+    call.updateSubscriptionsPartial(videoMode, {
+      [sessionId]: { dimension },
+    });
+
+    if (dimension) {
+      subscribedVideoLayoutRef.current = pendingVideoLayoutRef.current;
+      pendingVideoLayoutRef.current = undefined;
+    }
+  }, [
+    call,
+    isPublishingVideoTrack,
+    videoMode,
+    muteVideo,
+    sessionId,
+    hasJoinedCall,
+  ]);
+
+  useEffect(() => {
+    return () => {
+      subscribedVideoLayoutRef.current = undefined;
+      pendingVideoLayoutRef.current = undefined;
+    };
+  }, [videoMode, sessionId]);
+
+  const onLayout: React.ComponentProps<typeof RTCView>['onLayout'] = (
+    event,
+  ) => {
+    const dimension = {
+      width: Math.trunc(event.nativeEvent.layout.width),
+      height: Math.trunc(event.nativeEvent.layout.height),
+    };
+
+    // NOTE: If the participant hasn't published a video track yet,
+    // or the view is not viewable, we store the dimensions and handle it
+    // when the track is published or the video is enabled.
+    if (!call || !isPublishingVideoTrack || !hasJoinedCall) {
+      pendingVideoLayoutRef.current = dimension;
+      return;
+    }
+
+    // NOTE: We don't want to update the subscription if the dimension hasn't changed
+    if (
+      subscribedVideoLayoutRef.current?.width === dimension.width &&
+      subscribedVideoLayoutRef.current?.height === dimension.height
+    ) {
+      return;
+    }
+    call.updateSubscriptionsPartial(videoMode, {
+      [sessionId]: {
+        dimension,
+      },
+    });
+    subscribedVideoLayoutRef.current = dimension;
+    pendingVideoLayoutRef.current = undefined;
+  };
+
+  if (!canShowVideo) {
+    return (
+      <ParticipantVideoFallback participant={participant} onLayout={onLayout} />
+    );
+  }
+
+  return (
+    <RTCView
+      onLayout={onLayout}
+      streamURL={videoStreamToRender?.toURL()}
+      mirror={mirror}
+      style={styles.container}
+      objectFit={isScreenSharing ? 'contain' : 'cover'}
+      // zOrder should lower than the zOrder used in the floating LocalParticipantView
+      zOrder={Z_INDEX.IN_BACK}
+    />
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    ...StyleSheet.absoluteFillObject,
+  },
+});
