+++ packages/react-native-sdk/src/components/participants/index.ts
@@ -2,8 +2,8 @@ export * from './LocalParticipantView';
 export * from './ParticipantsInfoBadge';
 export * from './ParticipantLabel';
 export * from './ParticipantReaction';
-export * from './ParticipantVideo';
-export * from './ParticipantVideoPlaceholder';
+export * from './ParticipantVideoFallback';
 export * from './ParticipantView';
 export * from './ParticipantNetworkQualityIndicator';
 export * from './ParticipantsInfoList';
+export * from './VideoRenderer';
